+++ onnxruntime/core/providers/qnn/ort_api.h
@@ -23,8 +23,8 @@
 
 #include "onnxruntime_c_api.h"
 
-#include "core/providers/qnn/inlined_containers.h"
-#include "core/providers/qnn/int4.h"
+#include "core/providers/qnn/common/inlined_containers.h"
+#include "core/providers/qnn/common/int4.h"
 #include "onnxruntime_session_options_config_keys.h"
 #include "onnxruntime_run_options_config_keys.h"
