+++ onnxruntime/core/providers/qnn/qnn_execution_provider.cc
@@ -33,16 +33,15 @@
 #include "core/providers/qnn/builder/qnn_thread_pool.h"
 #include "core/providers/qnn/builder/qnn_utils.h"
 #include "core/providers/qnn/qnn_ep_utils.h"
-#include "core/providers/qnn/make_string.h"
 
 // Forward declarations for NodeUnit-related classes
 namespace onnxruntime {
 
 static std::string MakeSharedLibraryPath(std::string_view name) {
 #if defined(_WIN32)
-  return MakeString(name, ".dll");
+  return std::string(name) + ".dll";
 #else
-  return MakeString("lib", name, ".so");
+  return "lib" + std::string(name) + ".so";
 #endif
 }
